Move InterestingMemoryOperand from Transforms header to Analysis
